Ignore exit code of decoders on Windows
On Windows, some decoders run with errors. This seems to be due to
some memory issues according to
https://stackoverflow.com/questions/25987792/process-get-exited-with-return-value-3221226356

Most likely, the memory issues are also happening on Linux and macOS,
but Windows shows them up with an exit code != 0. We will investigate
further the potential memory problems with ASAN and/or Valgrind.