Add logic in Fluster to compare a reference raw video frame against another one
It's the class where we start to use unit tests, in this case, to compare synthetic images in RGBA and YUV420p format.
Now the .gitignore file includes the files used by pyCharm.
The makefile file has a new option, test, that scans the unit_test folder and launches all the tests contained in it.
The CI has a new step that launches the unit tests.
Added numpy module in Python requirements.txt file, it is needed for PSNR comparison