Run pylint with only 1 core
Running with -j0 is supposed to run with as many jobs as
cores available. However, it turns out that we tried it
and runs slower than with one single process. Probably
because the overhead of executing all processes in parallel
for a project the size of Fluster is not worth it.