Fix wrong report running parallel tests
We were reporting the test result in the worker. However,
we were collecting the results in the pool callback. This
created inconsitent results when using fail fast because
we had a report of a test being run ok but its result was
never collected because the callback was never called.

To fix it, we simply report the test result in the callback
to ensure consistency with what we collect.