OCP-2316 OCP-2351 (#87)
* OCP-2351 - Add field failing_test_vectors to test suite and json handlers

test_suite.py: necessary additions to accomodate for
failing_test_vectors within the .json test suite structure

test.py: same as above

* OCP-2316 - Add AAC conformance test suite

test_suites/aac/ISO_IEC_13818-4_2004.json
+ added test vectors from the equivalent ISO specification

ISO_IEC_14496-26_2010.json
+ added test vectors from the equivalent ISO specification

codec.py
+ added new pixel format called audio. now with the introduction of audio
we should discuss about refactoring/generalising this test suite filed

fdk_aac.py
+ refactoring after the introduction of static checks with black, mypy etc

* Simplify AAC tests logic

There's no need to use output_format in AAC test vectors.
Also, we can use a generic extension for the results output
file.

* Rename PixelFormat to a generic OutputFormat

Also move to_gst() method to GStreamer decoder file to
decouple it from the OutputFormat enum itself.

* Run pylint with only 1 core

Running with -j0 is supposed to run with as many jobs as
cores available. However, it turns out that we tried it
and runs slower than with one single process. Probably
because the overhead of executing all processes in parallel
for a project the size of Fluster is not worth it.

* Fix wrong report running parallel tests

We were reporting the test result in the worker. However,
we were collecting the results in the pool callback. This
created inconsitent results when using fail fast because
we had a report of a test being run ok but its result was
never collected because the callback was never called.

To fix it, we simply report the test result in the callback
to ensure consistency with what we collect.

* Fix total amount of tests to run in the report

* Add missing source checksums for AAC test suite

* Fix wrong alignment in test results report

* Added REFERENCE enum in TestVectorResult

This fixed a reporting bug introduced in a previous commit.

Co-authored-by: Pablo Marcos <pmarcos@fluendo.com>