Various GStreamer decoder fix and improvement (#111)
* gstreamer: Rename md5 parser function

This rename the function to be move accurate on what it does,
which is to parse the output of videocodectestsink element.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>

* gstreamer: Raise an exception if md5 could not be parsed

This improves the md5 parser so that an error is raised if
there was no MD5 found. This will allow an ERROR status rather
then failed status for the test.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>

* gstreamer: Check the process return code

When using videocodectestsink, check the process return code
and raise a CalledPRocessError allowing an ERROR status to
be traced instead of FAIL status. This fixes a small regression
recently introduced.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>

* gstreamer: Keep sterr inline with stout

When using write-less decoder in verbose mode, make sure to keep the
stderr output inline with the stout.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>

* Move running command with pipe into utils

Move GStreamer decoder code to run command and pipe the stdout into the
utils library so it can be used in other decoders later.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Co-authored-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>