Fallback to cwd as a base path for resources_dir & results_dir
Fluster defaults to using the script location as a base path for the
resources & results. This works fine for local runs of fluster, when
one is just calling fluster from a directory (e.g. ./fluster.py) but
when fluster is installed in a read-only system directory (e.g.
/usr/lib/python3.11/dist-packages/fluster for Debian) then the resources
and results cannot be written unless running as superuser.

Modify this behavior so that if the resources or results directories
cannot be found under the script path, use the current working directory
as a base path for these directories to allow fluster to be ran from a
system installation.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>