Using platform dirs to execute installed fluster.
Checking if fluster is uninstalled with the existence of the '.git'
dir.

For installed fluster, the site platform dirs is preferred for test
suites dir if exists. If not, user platform dirs will be used, that
is, ~/.local/share/fluster (or XDG_DATA_HOME) will be used if
/usr/share/fluster (or XDG_DATA_DIRS) doesn't exist. For resources dir
the user platform dirs is always used.

For uninstalled fluster, the usage of platform test suites dir [2] is
deprecated and it will be deleted in the future.

The platformdirs[1] lib was not used on purpose. The reason is to
avoid introducing the first external non-dev dependency.

[1] https://github.com/platformdirs/platformdirs
[2] https://github.com/fluendo/fluster/commit/a22f609