junitxml: Add each vector test time
In gitlab junit reports, the testsuite time is not used to show the
complete test time and is shown as 0.0 even if set.

Instead, it shows the sum of each test case time.

This adds each test case time in the junit report.
The sum will not give the actual total time when multiple threads are
used, but it can already give an idea of the used time.

Signed-off-by: Detlev Casanova <detlev.casanova@collabora.com>