feat: add TestVector optional decoder parameters
adding the capability to include an optional decoder parameters set
for a specfic test vector in the test suite definition.
optional parameters are utilized in Decoder class decode function abstraction
and are applied in Gstreamer(Decoder) and Ffmpeg(Decoder) classes implementations.
It's child class responsibility to handle the optional parameters.